rename component and move
