+++ ui/src/routes/_layout/chats/_layout-chats/$sessionId.tsx
@@ -39,7 +39,8 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { getSessionsQueryOptions } from "src/api/sessionApi.ts";
 import { getLlmModelsQueryOptions } from "src/api/modelsApi.ts";
-import { CustomErrorComponent } from "src/routes/_layout/chats/_layout-chats/index.tsx";
+
+import { CaiiTokenErrorComponent } from "src/components/ErrorComponents/CaiiTokenErrorComponent.tsx";
 
 export const Route = createFileRoute("/_layout/chats/_layout-chats/$sessionId")(
   {
@@ -48,6 +49,8 @@ export const Route = createFileRoute("/_layout/chats/_layout-chats/$sessionId")(
         context.queryClient.ensureQueryData(getSessionsQueryOptions),
         context.queryClient.ensureQueryData(getLlmModelsQueryOptions),
       ]),
-    errorComponent: (error) => <CustomErrorComponent errorComponent={error} />,
+    errorComponent: (errorComponent) => (
+      <CaiiTokenErrorComponent errorComponent={errorComponent} />
+    ),
   },
 );
